+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## HEAD
 
 - BREAKING: Do not automatically propagate `name` and `lastModified` in `file.slice()`. This matches the behavior of `File` more closely
+- In `slice(start, end)` make `end` default to `size` instead of `Infinity`. This more closely matches the `File` spec
 - Small perf improvement when streaming content arrays with Blobs in them and ending early
 
 ## v1.1.0 (Aug 22, 2024)